g2g: keep analytic_integral force kernels in scalar_type precision
Previously the generated qmmm/coulomb force kernels embedded literal
2.0/1.0 constants. Under -Wdouble-promotion these are double, so even
when the kernel is instantiated as <float> every multiply was promoted
to double, doubling register pressure and serialising the FP pipe on
Ampere. Wrap the literals in scalar_type(...) inside OS_expand.plx and
regenerate the 50 kernel headers; on fosfato (RTX 3080 Ti, aint_mp=1)
this drops QM/MM force gradients from ~250ms to ~30ms and the headline
qmmm_forces kernel from 246ms to 19ms.

Companion fixes that travel with the precision change:
  * os_util.h: replace the volatile-shared-memory warpReduce with
    __shfl_down_sync. Same FP reduction tree, but eliminates the
    intra-warp data race that compute-sanitizer racecheck flagged.
  * qmmm.cu: zero-initialise partial_mm_forces_dev so the
    COALESCED_DIMENSION padding columns don't trip initcheck on the
    cudaMemcpy in get_gradient_output.
  * coulomb.cu, os_common.cu, os_cutoff.cpp: explicit (double)/
    (scalar_type) casts at the boundaries the regenerated kernels now
    expose.

+++ g2g/analytic_integral/cuda/os_common.cu
@@ -318,9 +318,9 @@ void OSIntegral<scalar_type>::get_gradient_output(double* qm_forces, uint partia
     for (uint i = 0; i < G2G::fortran_vars.atoms; i++) {
       for (uint j = 0; j < partial_out_size; j++) {
         const auto& v = cpu_partial_qm_forces.data[i * w + j];
-        qm_forces[i + 0 * G2G::fortran_vars.atoms] += v.x;
-        qm_forces[i + 1 * G2G::fortran_vars.atoms] += v.y;
-        qm_forces[i + 2 * G2G::fortran_vars.atoms] += v.z;
+        qm_forces[i + 0 * G2G::fortran_vars.atoms] += (double)v.x;
+        qm_forces[i + 1 * G2G::fortran_vars.atoms] += (double)v.y;
+        qm_forces[i + 2 * G2G::fortran_vars.atoms] += (double)v.z;
       }
     }
